+++ lib/src/main/java/com/smileidentity/fragment/EnhancedSmartSelfieEnrollmentFragment.kt
@@ -7,25 +7,24 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.setFragmentResult
 import androidx.fragment.compose.content
 import com.smileidentity.SmileID
-import com.smileidentity.compose.SmartSelfieEnrollment
-import com.smileidentity.fragment.SmartSelfieEnrollmentFragment.Companion.KEY_REQUEST
-import com.smileidentity.fragment.SmartSelfieEnrollmentFragment.Companion.KEY_RESULT
-import com.smileidentity.fragment.SmartSelfieEnrollmentFragment.Companion.newInstance
-import com.smileidentity.fragment.SmartSelfieEnrollmentFragment.Companion.resultFromBundle
+import com.smileidentity.compose.SmartSelfieEnrollmentEnhanced
+import com.smileidentity.fragment.EnhancedSmartSelfieEnrollmentFragment.Companion.KEY_REQUEST
+import com.smileidentity.fragment.EnhancedSmartSelfieEnrollmentFragment.Companion.KEY_RESULT
+import com.smileidentity.fragment.EnhancedSmartSelfieEnrollmentFragment.Companion.newInstance
+import com.smileidentity.fragment.EnhancedSmartSelfieEnrollmentFragment.Companion.resultFromBundle
 import com.smileidentity.results.SmartSelfieResult
 import com.smileidentity.results.SmileIDResult
 import com.smileidentity.util.getParcelableCompat
-import com.smileidentity.util.randomJobId
 import com.smileidentity.util.randomUserId
 import com.squareup.moshi.Types
 import kotlinx.collections.immutable.toImmutableMap
 
 /**
- * Perform a SmartSelfie™ Enrollment
+ * Perform a Enhanced SmartSelfie™ Enrollment
  *
  * [Docs](https://docs.usesmileid.com/products/for-individuals-kyc/biometric-authentication)
  *
- * A [Fragment] wrapper for the [SmartSelfieEnrollment] to be used if not using Jetpack
+ * A [Fragment] wrapper for the [SmartSelfieEnrollmentEnhanced] to be used if not using Jetpack
  * Compose. New instances *must* be created via [newInstance]. Results are communicated back to the
  * caller via [setFragmentResult]. Therefore, the caller must use
  * [androidx.fragment.app.FragmentManager.setFragmentResultListener] to listen for the result. If
@@ -38,13 +37,13 @@ import kotlinx.collections.immutable.toImmutableMap
  *
  * Usage example:
  * ```java
- * SmartSelfieEnrollmentFragment smartSelfieFragment = SmartSelfieEnrollmentFragment
+ * EnhancedSmartSelfieEnrollmentFragment smartSelfieFragment = EnhancedSmartSelfieEnrollmentFragment
  *  .newInstance();
  * getSupportFragmentManager().setFragmentResultListener(
- *   SmartSelfieEnrollmentFragment.KEY_REQUEST,
+ *   EnhancedSmartSelfieEnrollmentFragment.KEY_REQUEST,
  *   this,
  *   (requestKey, result) -> {
- *     SmartSelfieResult smartSelfieResult = SmartSelfieEnrollmentFragment
+ *     SmartSelfieResult smartSelfieResult = EnhancedSmartSelfieEnrollmentFragment
  *       .resultFromBundle(result);
  *     getSupportFragmentManager()
  *       .beginTransaction()
@@ -54,29 +53,23 @@ import kotlinx.collections.immutable.toImmutableMap
  *   );
  * ```
  */
-class SmartSelfieEnrollmentFragment : Fragment() {
+class EnhancedSmartSelfieEnrollmentFragment : Fragment() {
     companion object {
-        const val KEY_REQUEST = "SmartSelfieEnrollmentRequest"
+        const val KEY_REQUEST = "EnhancedSmartSelfieEnrollmentRequest"
 
         /**
          * This is internal to prevent partners from accidentally using the wrong key. They only
          * need [KEY_REQUEST]. Partners should use [resultFromBundle] to extract the result.
          */
-        internal const val KEY_RESULT = "SmartSelfieEnrollmentResult"
+        internal const val KEY_RESULT = "EnhancedSmartSelfieEnrollmentRequest"
 
         /**
-         * Creates a new instance of [SmartSelfieEnrollmentFragment] which wraps the
-         * [SmileID.SmartSelfieEnrollment] Composable under the hood
+         * Creates a new instance of [EnhancedSmartSelfieEnrollmentFragment] which wraps the
+         * [SmileID.SmartSelfieEnrollmentEnhanced] Composable under the hood
          *
-         * @param userId The user ID to associate with the SmartSelfie™ Enrollment. Most often,
+         * @param userId The user ID to associate with the Enhanced SmartSelfie™ Enrollment. Most often,
          * this will correspond to a unique User ID within your own system. If not provided, a
          * random user ID will be generated.
-         * @param jobId The job ID to associate with the SmartSelfie™ Enrollment. Most often, this
-         * will correspond to a unique Job ID within your own system. If not provided, a random job ID
-         * will be generated.
-         * @param allowAgentMode Whether to allow Agent Mode or not. If allowed, a switch will be
-         * displayed allowing toggling between the back camera and front camera. If not allowed,
-         * only the front camera will be used.
          * @param showAttribution Whether to show the Smile ID attribution or not.
          * @param showInstructions Whether to deactivate capture screen's instructions for
          * @param skipApiSubmission Whether to skip the API submission and return the result of capture only
@@ -86,19 +79,15 @@ class SmartSelfieEnrollmentFragment : Fragment() {
         @JvmOverloads
         fun newInstance(
             userId: String = randomUserId(),
-            jobId: String = randomJobId(),
             allowNewEnroll: Boolean = false,
-            allowAgentMode: Boolean = false,
             showAttribution: Boolean = true,
             showInstructions: Boolean = true,
             skipApiSubmission: Boolean = false,
             extraPartnerParams: HashMap<String, String>? = null,
-        ) = SmartSelfieEnrollmentFragment().apply {
+        ) = EnhancedSmartSelfieEnrollmentFragment().apply {
             arguments = Bundle().apply {
                 this.userId = userId
-                this.jobId = jobId
                 this.allowNewEnroll = allowNewEnroll
-                this.allowAgentMode = allowAgentMode
                 this.showAttribution = showAttribution
                 this.showInstructions = showInstructions
                 this.skipApiSubmission = skipApiSubmission
@@ -116,11 +105,9 @@ class SmartSelfieEnrollmentFragment : Fragment() {
         savedInstanceState: Bundle?,
     ) = content {
         val args = requireArguments()
-        SmileID.SmartSelfieEnrollment(
+        SmileID.SmartSelfieEnrollmentEnhanced(
             userId = args.userId,
-            jobId = args.jobId,
             allowNewEnroll = args.allowNewEnroll,
-            allowAgentMode = args.allowAgentMode,
             showAttribution = args.showAttribution,
             showInstructions = args.showInstructions,
             skipApiSubmission = args.skipApiSubmission,
@@ -134,20 +121,10 @@ class SmartSelfieEnrollmentFragment : Fragment() {
 
 private val moshi = SmileID.moshi
 
-private const val KEY_ALLOW_AGENT_MODE = "allowAgentMode"
-private var Bundle.allowAgentMode: Boolean
-    get() = getBoolean(KEY_ALLOW_AGENT_MODE)
-    set(value) = putBoolean(KEY_ALLOW_AGENT_MODE, value)
-
 private const val KEY_USER_ID = "userId"
 private var Bundle.userId: String
     get() = getString(KEY_USER_ID)!!
     set(value) = putString(KEY_USER_ID, value)
-
-private const val KEY_JOB_ID = "jobId"
-private var Bundle.jobId: String
-    get() = getString(KEY_JOB_ID)!!
-    set(value) = putString(KEY_JOB_ID, value)
 
 private const val KEY_ALLOW_NEW_ENROLL = "allowNewEnroll"
 private var Bundle.allowNewEnroll: Boolean